app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.52.0
databricks-sdk>=0.30.0
pydantic>=2.0.0
mlflow>=3.10.1
asyncpg>=0.29.0
databricks-vectorsearch>=0.40
